+++ src/pages/collection/[id].tsx
@@ -1,18 +1,35 @@
-import { FileStack } from "lucide-react";
 import { GetServerSideProps, InferGetStaticPropsType } from "next";
 
 import { ApiClient } from "@/api/http-common";
-import { BreadCrumbs } from "@/components/breadcrumbs/Breadcrumbs";
+import { Columns } from "@/components/atoms/columns/Columns";
 import Layout from "@/components/layouts/Main";
-import { SubNav } from "@/components/nav/SubNav";
-import { MultiCol } from "@/components/panels/MultiCol";
-import { SingleCol } from "@/components/panels/SingleCol";
-import { Heading } from "@/components/typography/Heading";
+import { ContentsSideBar } from "@/components/organisms/contentsSideBar/ContentsSideBar";
 import { withEnvConfig } from "@/context/EnvConfig";
-import { TCollection, TTheme } from "@/types";
+import { TCollection, TTheme, TThemeConfig } from "@/types";
 import { getFeatureFlags } from "@/utils/featureFlags";
 import { isLitigationEnabled } from "@/utils/features";
 import { readConfigFile } from "@/utils/readConfigFile";
+
+interface IProps {
+  collection: TCollection;
+  theme: TTheme;
+  themeConfig: TThemeConfig;
+}
+
+const CollectionPage: InferGetStaticPropsType<typeof getServerSideProps> = ({ collection, theme, themeConfig }: IProps) => {
+  return (
+    <Layout title={collection.title} description={collection.description} theme={theme} themeConfig={themeConfig} metadataKey="collection">
+      <Columns>
+        <ContentsSideBar items={[]} stickyClasses="!top-[72px] pt-3 cols-2:pt-6 cols-3:pt-8" />
+        <main className="py-3 cols-2:py-6 cols-3:py-8 cols-3:col-span-2 cols-4:col-span-3">
+          <pre className="w-full max-h-[700px] bg-surface-ui text-sm text-text-tertiary overflow-scroll">{JSON.stringify(collection, null, 2)}</pre>
+        </main>
+      </Columns>
+    </Layout>
+  );
+};
+
+export default CollectionPage;
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   context.res.setHeader("Cache-Control", "public, max-age=3600, immutable");
@@ -45,59 +62,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
     props: withEnvConfig({
       collection,
       theme,
+      themeConfig,
     }),
   };
 };
-
-type TProps = {
-  collection: TCollection;
-  theme: TTheme;
-};
-
-const CollectionPage: InferGetStaticPropsType<typeof getServerSideProps> = ({ collection, theme }: TProps) => {
-  const { families } = collection;
-  const breadcrumbCategory = { label: "Search results", href: "/search" };
-
-  return (
-    <Layout title={collection.title} description={collection.description} theme={theme}>
-      <section className="mb-8">
-        <SubNav>
-          <BreadCrumbs category={breadcrumbCategory} label={collection.title} />
-        </SubNav>
-        <MultiCol>
-          <SingleCol extraClasses={`mt-8 px-5 w-full`}>
-            <Heading level={1}>{collection.title}</Heading>
-            <div className="flex flex-wrap text-sm gap-1 mt-2 items-center middot-between" data-cy="family-metadata">
-              Collection
-            </div>
-            <section className="mt-6">
-              <div className="text-content">{collection.description}</div>
-            </section>
-
-            {families.length > 0 && (
-              <section className="mt-10">
-                <Heading level={2}>Documents</Heading>
-                {families.map((family) => (
-                  <a
-                    key={family.slug}
-                    href={`/document/${family.slug}`}
-                    className="family-document group mt-4 p-4 rounded-lg border bg-gray-25 border-gray-100 shadow-xs transition duration-300 flex flex-no-wrap cursor-pointer hover:border-blue-100 hover:bg-gray-50"
-                  >
-                    <div className="flex-0 mr-2 hidden md:block">
-                      <FileStack size="20" className="mt-0.5 text-text-brand" />
-                    </div>
-                    <div className="flex-1">
-                      <span dangerouslySetInnerHTML={{ __html: family.title }} />
-                    </div>
-                  </a>
-                ))}
-              </section>
-            )}
-          </SingleCol>
-        </MultiCol>
-      </section>
-    </Layout>
-  );
-};
-
-export default CollectionPage;
